File cleanup; Moves all classes in their own files and updates build files
