+++ bark_ml/environments/single_agent_runtime.py
@@ -57,8 +57,7 @@ class SingleAgentRuntime(Runtime):
     # set actions
     eval_id = self._scenario._eval_agent_ids[0]
     if eval_id in self._world.agents:
-      print(f'action: {action}')
-      self._world.agents[eval_id].behavior_model.ActionToBehavior(action)
+      self._world.agents[eval_id].behavior_model.ActionToBehavior(action[0])
 
     # step and observe
     self._world.Step(self._step_time)
