+++ bark_ml/library_wrappers/lib_tf2_gnn/BUILD
@@ -1,17 +1,13 @@
 py_library(
-    name = "gnn_wrapper",
-    srcs = ["gnn_wrapper.py"],
+    name = "lib_tf2_gnn",
+    srcs = [
+        "__init__.py",
+        "gnn_wrapper.py", 
+        "gnn_actor_network.py",
+        "gnn_critic_network.py"
+    ],
     data = ["@bark_project//python:bark.so"],
     imports = ["../external/bark_project/python/"],
     deps = ["@bark_project//modules/runtime:runtime"],
-    visibility = ["//visibility:public"],
-)
-
-py_library(
-    name = "gnn_actor_network",
-    srcs = ["gnn_actor_network.py"],
-    data = ["@bark_project//python:bark.so"],
-    imports = ["../external/bark_project/python/"],
-    deps = [],
     visibility = ["//visibility:public"],
 )
