+++ bark_ml/library_wrappers/lib_tf_agents/BUILD
@@ -1,21 +1,7 @@
 py_library(
     name = "tfa_wrapper",
     srcs = [
-        "tfa_wrapper.py",
-        "gnn_wrapper.py", 
-        "gnn_actor_network.py"
-    ],
-    data = ["@bark_project//python:bark.so"],
-    imports = ["../external/bark_project/python/"],
-    deps = ["@bark_project//modules/runtime:runtime"],
-    visibility = ["//visibility:public"],
-)
-
-py_library(
-    name = "gnn",
-    srcs = [
-        "gnn_wrapper.py", 
-        "gnn_actor_network.py"
+        "tfa_wrapper.py"
     ],
     data = ["@bark_project//python:bark.so"],
     imports = ["../external/bark_project/python/"],
