+++ examples/BUILD
@@ -55,6 +55,7 @@ py_test(
   deps = [
     "//bark_ml/environments:single_agent_runtime",
     "//bark_ml/environments:blueprints",
+    "//bark_ml/library_wrappers/lib_tf2_gnn:lib_tf2_gnn",
     "//bark_ml/library_wrappers/lib_tf_agents/agents:agents",
     "//bark_ml/library_wrappers/lib_tf_agents/runners:runners",
   ]
